Update api.js
Written the logic for the .get and .delete methods. Both work correctly. Now when we go to the /api/issues/apitest path we get all the issue entries from the database for the apitest project. From there more issues can be added or they can be deleted. Next step is to code the .put method logic for updating the issues or closing them.