Update api.js
Updated the .get() method logic to allow for query string filtering based on the req.query obj. Optimized the .post() method with a standardIssue object as a model for the new issues. Replaced res.json() method with .send() to make the assertions from the tests work as expected. Written the .put() method and works as it should when updating or closing an issue. Once closed an issue cannot be opened, just deteled. All 11 functional tests are passing. The project is now complete and ready to be submitted to FCC for review.